CRUD de la tabla pedido cliente completo y funcional

+++ src/main/java/com/TicoTaco/TicoTaco/Repository/CajasRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CajasRepository extends JpaRepository<CajasModel, Integer> {
+public interface CajasRepository extends JpaRepository<CajasModel, Long> {
 }
